Frontend: Refactor type SceneMapRouteProps to types.ts
